+++ examples/shopping_app/lib/core/db/shopping_cache_database.g.dart
@@ -34,6 +34,9 @@ final class _$ShoppingCacheDatabase implements ShoppingCacheDatabase {
   Connection get connection => _driver;
 
   @override
+  Future<Result<Unit, SqlxError>> migrate() => Future<Result<Unit, SqlxError>>.value(const Ok(unit));
+
+  @override
   UnsafeSql get unsafe => Sqlite3UnsafeSql(_driver);
 
   Pool get pool => _driver;