+++ fixtures/postgres_app/lib/postgres_app.dart
@@ -1,0 +1,6 @@
+/// PostgreSQL Database fixture.
+library;
+
+export 'src/database.dart';
+export 'src/order.dart';
+export 'src/orders_repo.dart';